Added `-cs` target

+++ tool/src/massive/munit/command/GenerateCommand.hx
@@ -38,51 +38,37 @@ class GenerateCommand extends MUnitCommand
 	private var includeHxml:Bool;
 	private var testFilter:String;
 	
-	public function new():Void
+	public function new()
 	{
 		super();
 		testFilter = null;
 		includeHxml = true;
 	}
 	
-	override public function initialise():Void
+	override public function initialise()
 	{
 		var testSrcPath = console.getNextArg();
 		dir = initialiseTestSourceDirectory(testSrcPath);
 		
 		includeHxml = console.getOption("nohxml") != "true";
-		
 		if(!includeHxml)
 		{
 			var hxmlPath  = console.getNextArg();
 			hxmlOutput = initialiseHxmlOutputFile(hxmlPath);
 		}
-	
-		
-		
 		testFilter = console.getOption("-filter");
 	}
 
-	
-	private function initialiseTestSourceDirectory(?testSrcPath:String=null):File
+	function initialiseTestSourceDirectory(?testSrcPath:String):File
 	{
 		var dir:File;
 		
 		if(testSrcPath == null)
 		{
 			dir = config.src;
-			if(dir == null)
-			{
-				error("Default test src directory is not set. Please run munit config.");
-			}
-			
-			
-			if(!dir.exists)
-			{
-				error("Default test src directory does not exist (" + config.dir.getRelativePath(dir) + ").\nPlease run munit config or specify a <src> and <hxml> path");
-			}
+			if(dir == null) error("Default test src directory is not set. Please run munit config.");
+			if(!dir.exists) error("Default test src directory does not exist (" + config.dir.getRelativePath(dir) + ").\nPlease run munit config or specify a <src> and <hxml> path");
 		}
-		
 		else
 		{
 			dir = File.create(testSrcPath, console.dir);
@@ -98,64 +84,49 @@ class GenerateCommand extends MUnitCommand
 				}
 			}
 		}
-		
 		return dir;
 	}
 
-	
-	private function initialiseHxmlOutputFile(?hxmlPath:String=null):File
+	function initialiseHxmlOutputFile(?hxmlPath:String):File
 	{
 		var hxmlOutput:File = null;
-		
 		if(hxmlPath == null) 
 		{
-			if( config.hxml == null)	error("Default hxml path is not set. Please run munit config or specify a hxml file path.");
+			if(config.hxml == null) error("Default hxml path is not set. Please run munit config or specify a hxml file path.");
 			else hxmlOutput =  config.hxml;
 		}
-		else
-		{
-			hxmlOutput = File.create(hxmlPath, config.dir); 
-		}
-		
-		
-		
+		else hxmlOutput = File.create(hxmlPath, config.dir); 
 		if(hxmlOutput == null || !hxmlOutput.isFile) error("Invalid hxml path " + hxmlPath);
-		
-		
 		return hxmlOutput;
 	}
 	
 
-	override public function execute():Void
+	override public function execute()
 	{
 		if(!hasTestMain())
 		{
 			creatTestMainClass();
 			createExampleTestClass();
-			
 			if(includeHxml) createTestHxmlFile();
-		
 		}
-		
 		createTestSuiteClass();
 	}
 	
 	
-	private function hasTestMain():Bool
+	function hasTestMain()
 	{
 		var testMain:File = dir.resolvePath("TestMain.hx");
 		return testMain.exists;
 	}
-
 	
-	private function creatTestMainClass():Void
+	function creatTestMainClass()
 	{
 		var testMain:File = dir.resolvePath("TestMain.hx");
 		var content = TemplateUtil.getTemplate("test-main", {url:HTTPClient.DEFAULT_SERVER_URL});
 		testMain.writeString(content, true);
 	}
 	
-	private function createExampleTestClass():Void
+	function createExampleTestClass()
 	{
 		//create an example test class for reference
 		var testExample:File = dir.resolvePath("ExampleTest.hx");
@@ -163,64 +134,50 @@ class GenerateCommand extends MUnitCommand
 		testExample.writeString(content, true);
 	}
 	
-	private function createTestHxmlFile():Void
+	function createTestHxmlFile()
 	{
-		if(hxmlOutput == null) return;
-		if(hxmlOutput.exists) return;
-			
+		if(hxmlOutput == null || hxmlOutput.exists) return;
 		var src:String = config.src != null ? config.dir.getRelativePath(config.src) + "" : "";
 		var bin:String = config.bin != null ? config.dir.getRelativePath(config.bin) + "" : "";
-
 		//write out stub hxml file
 		var content = TemplateUtil.getTemplate("test-hxml", {src:src, bin:bin});
 		hxmlOutput.writeString(content, true);
 	}
 	
-	
-	private function createTestSuiteClass():Void
+	function createTestSuiteClass()
 	{
 		var classes:Array<String> = getFilteredClassesInDirectory(dir);
 		var content:String = generateTestSuiteClassFromClasses(classes);
-
 		var testSuite:File = dir.resolvePath("TestSuite.hx");
 		testSuite.writeString(content, true);
-	}	
+	}
 	
-	private function getFilteredClassesInDirectory(dir:File):Array<String>
+	function getFilteredClassesInDirectory(dir:File):Array<String>
 	{
-	
 		var files:Array<File> = dir.getRecursiveDirectoryListing(~/.*Test\.hx$/);
-	
 		var classes:Array<String> = [];
 		var clasz:String;
-
 		for(file in files)
 		{
 			clasz = dir.getRelativePath(file).substr(0, -3);
 			clasz = clasz.split("/").join(".");
 			if(clasz == "TestMain") continue;
 			if(testFilter != null && clasz.indexOf(testFilter) == -1) continue;
-			
 			classes.push(clasz);
-			
-		}	
 		
+		}
 		return classes;
 	}
 	
-	private function generateTestSuiteClassFromClasses(classes:Array<String>):String
+	function generateTestSuiteClassFromClasses(classes:Array<String>):String
 	{
 		var imports:String = "";
 		var tests:String = "";
-		
 		for(clasz in classes)
 		{
 			imports += "\nimport " + clasz + ";";
 			tests += "\n		add(" + clasz + ");";
 		}
-		
 		return TemplateUtil.getTemplate("test-suite", {imports:imports,tests:tests});
 	}
-	
-
 }
